Reworked parts of the build system
1) Image creation now works as the user without root priviledges
2) Enabled SMP for something later on :)
3) Reworked .clang-format rules again
4) The 'all' targen on the Makefile also creates the image
5) Various other bits and bobs

+++ scripts/create_img.sh
@@ -7,23 +7,40 @@ die() {
     exit 1
 }
 
-if [ $(/usr/bin/id -u) != "0" ]; then
-    exec sudo -E -- "$0" "$@" || die "this script needs root priviledges"
-fi
-
 safe_exit() {
-    if [ -d "$BIN_DIR/mnt" ]; then
-        rm -rf mnt || ( umount mnt && rm -rf mnt )
+    if [ -d "$BIN_DIR"/mnt ]; then
+        if kill -0 "$pid"; then
+            guestunmount "$BIN_DIR"/mnt
+	fi
     fi
 }
 
 trap safe_exit EXIT
 
-qemu-img create "$BIN_DIR/img" 100M
-mkfs.ext2 -q "$BIN_DIR/img"
-mkdir -p "$BIN_DIR/mnt"
-mount "$BIN_DIR/img" "$BIN_DIR/mnt"
-cp -r "$BIN_DIR"/sysroot/* "$BIN_DIR/mnt"
-umount "$BIN_DIR/mnt"
-rm -rf "$BIN_DIR/mnt"
-chmod 0666 "$BIN_DIR/img"
+if [ ! -e "$BIN_DIR"/img ]; then
+    qemu-img create "$BIN_DIR"/img 100M
+    mkfs.ext2 -q "$BIN_DIR"/img
+fi
+
+mkdir -p "$BIN_DIR"/mnt
+
+guestmount -a "$BIN_DIR"/img -m /dev/sda --pid-file mount.pid "$BIN_DIR"/mnt
+pid="$(cat mount.pid)"
+rm -rf "${BIN_DIR:?}"/mnt/*
+cp -r "$BIN_DIR"/sysroot/* "$BIN_DIR"/mnt
+guestunmount "$BIN_DIR"/mnt
+
+timeout=10
+
+count=$timeout
+while kill -0 "$pid" 2>/dev/null && [ $count -gt 0 ]; do
+    sleep 1
+    count=$((count+1))
+done
+
+if [ $count -eq 0 ]; then
+    echo "$0: wait for guestmount to exit failed after $timeout seconds"
+    exit 1
+fi
+
+chmod 0666 "$BIN_DIR"/img
